keyword cloud - 키워드텍스트 및 버튼 위젯을 담당하는 위젯을 분리
